requirements.txt
google-generativeai
python-dotenv
streamlit
stellar-sdk